requirements.txt
albumentations
numpy
opencv-python
Pillow
scikit-image
torch
torchvision
torchmetrics